server/list_models.py
import os
from dotenv import load_dotenv
import google.generativeai as genai

load_dotenv()
api_key = os.getenv("GEMINI_API_KEY")
if not api_key:
    raise ValueError("GEMINI_API_KEY missing")

genai.configure(api_key=api_key)

models = genai.list_models()
print("Available models:")
for m in models:
    print(m.name)
